testing/backendapp/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.3
aioshutil==1.5
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.3.0
apify==1.7.1
apify_client==1.6.4
apify_shared==1.1.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Authlib==1.3.1
autobahn==24.4.2
Automat==24.8.1
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
Babel==2.15.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
boto3==1.36.26
botocore==1.36.26
Brlapi==0.8.3
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
channels==4.2.0
channels_redis==4.2.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
constantly==23.10.4
cryptography==44.0.1
cupshelpers==1.0
Cython==3.0.11
daphne==4.1.2
dateparser==1.2.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
email_validator==2.1.1
eventkit==1.0.3
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.109.2
fasteners==0.14.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
fqdn==1.5.1
frozenlist==1.5.0
future==0.18.2
fyers_apiv3==3.1.5
gunicorn==22.0.0
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hyperlink==21.0.0
ib-insync==0.9.86
idna==3.7
importlib-metadata==4.6.4
incremental==24.7.2
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
mangum==0.17.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.0
protobuf==3.12.4
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycryptodome==3.21.0
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.1.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.12.1
pytz==2025.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rjsmin==1.2.3
rpds-py==0.18.1
s3transfer==0.11.2
scipy==1.15.1
SecretStorage==3.3.1
Send2Trash==1.8.3
service-identity==24.2.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.36.3
systemd-python==234
terminado==0.18.1
texttable==1.6.4
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2025.1
tzlocal==5.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.30.5
virtualenv==20.26.3
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
Werkzeug==3.0.3
whitenoise==6.9.0
widgetsnbextension==4.0.11
wrapt==1.17.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
zope.interface==7.2
